% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{get_polygons}
\alias{get_polygons}
\title{Get polygons from voronoi map}
\usage{
get_polygons(tass)
}
\arguments{
\item{tass}{Voronoi Map}
}
\value{
data frame of polygons
}
\description{
Get polygons from voronoi map
}
\examples{
get_polygons(voronoi_map(1:3))
}
