% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plot.voronoi_map}
\alias{plot.voronoi_map}
\title{Plotting method for voronoi maps}
\usage{
\method{plot}{voronoi_map}(x, y, ...)
}
\arguments{
\item{x}{A valid voronoi_map object}

\item{y}{Ignored}

\item{...}{Ignored}
}
\value{
Doesn't return anything; draws plot as side effect
}
\description{
Plotting method for voronoi maps
}
\examples{
plot.voronoi_map(obj)
}
